+++ frontend/src/components/Header/Header.jsx
@@ -1,4 +1,4 @@
-import { Navigation } from "../Navigation/Navigation";
+import { HeaderNavigation } from "../HeaderNavigation/HeaderNavigation";
 import { UserMenu } from "../UserMenu/UserMenu";
 import { HeaderWrapper, HeaderContainer } from "./Header.styled";
 
@@ -7,7 +7,7 @@ export const Header = () => {
     <>
       <HeaderWrapper>
         <HeaderContainer>
-          <Navigation />
+          <HeaderNavigation />
           <UserMenu />
         </HeaderContainer>
       </HeaderWrapper>
