+++ frontend/src/components/HeaderNavigation/index.js
@@ -1,0 +1,1 @@
+export * from "./HeaderNavigation";
