--- frontend/src/components/Navigation/Navigation.jsx
@@ -1,37 +1,0 @@
-import { Link, Wrapper, LogoTitle } from "./Navigation.styled";
-
-export const Navigation = () => {
-  return (
-    <nav>
-      <Link to="/">
-        <Wrapper>
-          <svg
-            width="30"
-            height="30"
-            viewBox="0 0 30 30"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-          >
-            <path
-              d="M3.52059 7.14286L18.7187 2.88927L17.9139 1.22119C17.3873 0.13694 16.1029 -0.310807 15.0455 0.229123L1.43994 7.14286H3.52059Z"
-              fill="#24CCA7"
-            />
-            <path
-              d="M23.2625 2.85742C23.0793 2.85742 22.8962 2.88475 22.713 2.9394L19.2414 3.97781L8.6394 7.14314H20.6734H25.9182L25.2688 4.53801C25.0191 3.52237 24.1782 2.85742 23.2625 2.85742Z"
-              fill="#24CCA7"
-            />
-            <path
-              d="M26.5873 8.03711H26.2032H25.6811H25.1589H21.1735H6.52891H4.60865H2.99158H2.69175H1.68782C1.15553 8.03711 0.680516 8.2811 0.370578 8.66548C0.229085 8.84263 0.12128 9.04651 0.0606401 9.27045C0.0235822 9.41083 0 9.5579 0 9.70831V9.90885V11.814V28.3288C0 29.2513 0.754632 30 1.68445 30H26.5839C27.5137 30 28.2684 29.2513 28.2684 28.3288V23.6662H18.2729C16.6929 23.6662 15.4093 22.3927 15.4093 20.8251V19.2943V18.7762V18.2582V17.1084C15.4093 16.3396 15.7193 15.6411 16.2212 15.1297C16.6659 14.6751 17.2622 14.3676 17.9292 14.2907C18.0404 14.2774 18.155 14.2707 18.2695 14.2707H26.8669H27.3891H27.9113H28.2684V9.70831C28.2717 8.78581 27.5171 8.03711 26.5873 8.03711Z"
-              fill="#4A56E2"
-            />
-            <path
-              d="M29.4508 15.8685C29.2824 15.7148 29.0836 15.5978 28.8613 15.5209C28.6894 15.4641 28.5075 15.4307 28.3155 15.4307H28.2717H28.238H27.7158H25.8326H18.2728C17.343 15.4307 16.5884 16.1794 16.5884 17.1019V17.9341V18.4522V18.9703V20.8219C16.5884 21.7444 17.343 22.4931 18.2728 22.4931H28.2717H28.3155C28.5075 22.4931 28.6894 22.4597 28.8613 22.4029C29.0836 22.3294 29.2824 22.209 29.4508 22.0553C29.7877 21.7511 29.9999 21.3099 29.9999 20.8219V17.1019C29.9999 16.6139 29.7877 16.1727 29.4508 15.8685ZM21.7899 19.2945C21.7899 19.7557 21.4126 20.1301 20.9477 20.1301H20.3885C19.9236 20.1301 19.5463 19.7557 19.5463 19.2945V18.7396C19.5463 18.4722 19.6709 18.2349 19.8697 18.0845C20.0145 17.9742 20.1931 17.904 20.3885 17.904H20.53H20.9477C21.4126 17.904 21.7899 18.2784 21.7899 18.7396V19.2945Z"
-              fill="#4A56E2"
-            />
-          </svg>
-          <LogoTitle>Wallet</LogoTitle>
-        </Wrapper>
-      </Link>
-    </nav>
-  );
-};
