--- frontend/src/components/Navigation/index.js
@@ -1,1 +1,0 @@
-export * from "./Navigation";
